Use groupId.artifactId on apklib name clash
This commit keeps the old artifactId-only nomenclature for apklibs
(retaining backward compatibility with already existing projects)
and switches to groupId.artifactId nomenclature when there is a
name clash between two (or more) referenced apklibs.

The code uses an intermediate HashMap to avoid scanning the apklib
list for duplicates each time a new library dependency is found.