[task #155] Added Non-Runtime Classpath Container
Added separate classpath container for non-runtime Maven dependencies. This was
created for two reasons:

  1. Allow test-scope dependencies to remain in the Eclipse workspace
  classpath, which means we can successfully run the default Eclipse JUnit test
  runner for Android/Maven tests and remove the need for separate test projects
  to run unit tests.

  2. Due to the abuse of the "provided" Maven dependency-scope, some
  dependencies (such as Lombok) are added to this scope but aren't actually
  provided by the platform. These still need to be added to the workspace
  classpath despite being labelled "provided".

The creation of this classpath container means we can still get the advantages
of adding non-runtime dependencies to Eclipse project workspaces without the
disadvantage of having unwanted dependencies included in the APK build step.
Previously we were stripping all non-runtime dependencies (e.g. test, provided,
etc.) from the classpath container but this was too broad.

We now have to explicitly strip the actual platform-provided dependencies
from both the Maven and Non-Runtime classpath container. This looks for the
dependency with the artifact ID "android" and group ID "com.google.android" and
strips it, and its dependencies, from all classpaths. Obviously, if these magic
strings change this will break this code so its more brittle than the old
method.

This change includes a full suite of integration tests to verify the new
classpath container behaves as expected.

There maybe some problems associated with this change: for example, the new
classpath container may break other plugins that expect all Maven dependencies
to be in the one, Maven classpath container.

Several re-factorings were also made as part of this change:

The project configuration code (modifying the actually project parameters) was
renamed to the more accurate "workspace configuration" code.

I refactored the classpath configuration code into separate "Command" design
patterns (as has been previously done for the workspace configuration code),
increasing the readability of the code and removing the "god object" that was
responsible for configuring the Eclipse workspace classpath.

Lastly, some basic logging was added to the application. This logging maybe
overly verbose for a production release.