Re-factored classpath configuration code
Re-factored classpath configuration code to isolate Eclipse-specific
code behind the new adapter classes MavenEclipseClasspath and
EclipseSourceEntry. This work is only partially done as the abstraction
is still leaking Eclipse details (for example, the Eclipse container
IDs).

Also removed the Project class which was a bad abstraction, and merged
this code into the AdtEclipseAndroidProject class.

This change has meant we aren't creating instances of the
AdtEclipseAndroidProject in its factory for the case of classpath
configuration. This should be fixed in later re-factorings, perhaps
replacing the factory with a builder design pattern.

This re-factoring should make it easier to code for classpath changes,
such as for modifying the output locations for test source directories
(enhancement #123) or retaining modifications to the .classpath file
(bug #182).