Removed AndroidProject super-interface
Deleted AndroidProject super-interface from code. This abstraction
created a common interface for Jawyay Maven android plugin projects
(which we only want to read values from to modify the Eclipse workspace)
and ADT android projects (which we want to read and modify).

These two classes actually unrelated behaviour and a common interface
gives the false impression they can be substituted for each other.

This re-factoring has the benefit of removing several unsupported
operations from classes as well as removing redundant code that is never
called. It also makes the code more type-safe ensuring we can't write
values to the Maven POM.