Add platform deps and binaries to JUnit classpath
In order to enable the execution of unit tests using frameworks like
Robolectric, the classpath of the JUnit launcher has been modified to
include platform provided dependencies and the binary classses.

This applies only to Android and Maven projects. This relies on a new
classpath provided being applied to the launcher configuration. This
config is applied whenever a launcher is modified, or the Maven nature
is added to a project.

Note: due to limitations in the Eclipse PDE unit testing framework this
behaviour could not be verified. A test for this has been added but
quarantined. Test is successful in isolation but failing when run with
other fixtures probably due to some kind of timing issue.

Tests were written to verify Android binary classes folder and platform
dependencies have been added to Junit launcher classpath.

This fixes a number of outstanding issues:

https://github.com/rgladwell/m2e-android/issues/52
https://github.com/rgladwell/m2e-android/issues/154

As well as being the last fix for full unit-test support:

https://github.com/rgladwell/m2e-android/issues/15