Fixed regression introduced by Kepler SR2
Eclipse Kepler SR2 introduced a regression in the Guice API bundled with
it. Guice no longer recognises the javax.inject.Inject annotation. This
was resolved by replacing the annotation with the Guice @Inject
annotation.

However, this makes the code less re-usable and creates a tighter
dependency on the Guice API. This should be reported upstream for a fix.

https://github.com/rgladwell/m2e-android/issues/187