Added repository to archetype catalog
Creating a project from archetype via the New Project wizard failed
with error: The desired archetype does not exist.
This commit fixes the problem by adding repository url for each
archetype, so maven searches the central repository for them instead
of our repository.
The version numbers of the archetypes are also updated.