Fixed library project dependencies on different device
ADT assumes that the dependency library project is in a relative path
to the main project. If the library project is in a different device,
it cannot be refered relatively. Since we do support library project
dependencies outside of the workspace, this problem is real.

This commit reimplements ProjectState.needs(ProjectState) method to fix
this issue, utilizing reflection to call private methods.