Fixed build error on unregistered Maven project.
Build errors were being caused by an unhandled NPE exception in the code
to prune platform-provided dependencies from the non-runtime Maven
classpath container. This was due to the fact the Maven project had not
yet registered.

Part of this change also included adding basic logging.

This fixes:

https://github.com/rgladwell/m2e-android/issues/198

It also seems to resolve the issue where opening and closing the project
will create duplicate projects in the Eclipse Package Explorer:

https://github.com/rgladwell/m2e-android/issues/197