Attach sources and JavaDoc to NonRuntime container
The entries in the Non-Runtime Maven Dependencies classpath container
does not have sources or JavaDoc attached automatically from maven
repositories. If the user tries to attach manually, it does not gets
updated.

This commit addresses this issue: the artifacts are attached
automatically if m2e is configured to do so, also manual attachment is
supported. The codebase is adapted from the upstream m2e-jdt project,
because these parts are not part of the public API.

Only schedule a download when the currently set source or JavaDoc
attachment is null. This prevents m2e-a from re-scheduling downloads
forever if the artifact (sources, Javadoc) is not available.

https://github.com/rgladwell/m2e-android/issues/188