Adding provided workspace dependency projects to NonRuntime container
A quite common use-case is keeping both the application project and the
test project in the same workspace, where the app project is a provided
dependency of the test project. In this case m2e workspace resolution is
triggered, which adds the project as a classpath entry instead of the
jar library. The current code does not aware of this situation and keeps
the project entry in the Runtime container.

This commit addresses the issue, by adding the project entry the the
NonRuntime container and removing it from the Runtime one, but only if
the discussed m2e feature is in use and the dependency project is open.