Support custom path for AndroidManifest.xml and res folder
The maven-android-plugin enables the user to define custom paths for
AndroidManifest.xml and res folder. Our plugin did not take into account
these properties. This commit addresses the issue by reading the
parameters from the POM and linking the AndroidManifest.xml file and the
res folder to the custom paths.