Added backwards compatible support for Luna and m2e 1.5.
Added support for the m2e 1.5 and backwards compatability for older
versions such as 1.4.

This change is required due to the breaking package name change for the
Aether API introduced when Aether was moved to the Eclipse foundation:

https://projects.eclipse.org/projects/technology.aether/releases/0.9.0/plan

Unfortunately, the pre-1.5 m2e extension APIs are built directly against
the Aether API requiring we support both in the same code-base.

http://github.com/rgladwell/m2e-android/issues/issue/220

When users stop using the Android depednecies in Maven Central we can
retire all resolution code.