Issue #236 : CLI Tests Fail when android sdk 10 and 7 missing
This fixes a configuration issue with Tycho so that the complete Eclipse IDE
is included as part of the test runtime.  This will now correctly fail the
9 tests that depend on a valid project configuration for android 10 and
android 7, but allow the other tests to pass.

The issue was that not all the required items that ADT needed were being
installed.  All target platforms are updated with the appropriate entry to
install the necessary feature, as well as the tests depend on the feature
being installed.