Ensure that source and test folders are sorted.
After running the other Classpath Configuration classes
the source folders are out of sequence, causing all sorts
of problems looking up source. This is an attempt to
re-order the classpaths.

The order of the entries are:

src/main/java
src/main/resources
src/test/java
src/test/resources
Android Framework
Android Private Libraries
Android Dependencies
Maven Dependencies
Non-runtime Maven Dependencies

All other entries are kept the same in the classpath.  The
reordering has to be the last entry to run in the
RawClasspathConfigurations.