Add test launch config after project imported
.launch files are read before the project is fully configured by
m2e-android. m2e-android only modifies the launch config if the project
has Android nature. But when .launch is read at import, the project has
no Android nature, yet. So in the testcase, we have to add the .launch
configuration file after the project is configured (so Android nature is
added).

This way we can reenable
"testConfigureAddsAndroidTestClasspathProviderToTestRunner" testcase.
Also now we can successfully run a test project which actually tests
a class from the application, nut just uses the test class itself. This
was not possible before, because our test classpathprovider was not
added to the launch configuration at project import.