Hardcode platform provided dependencies
We remove the dependencies (and transitive dependencies) of the Android
artifact from the Maven classpath container, since the Android library
entry already contains these classes. Currently we do with this
dynamically with Aether, resolving all the dependencies and removing
them. Unfortunately we have problems with Aether. This commit replaces
resolution with Aether by simply returning the hardcoded dependency
lists for the artifacts.

https://github.com/rgladwell/m2e-android/issues/259