Generate project properties from POM
The project.properties file determines which Android target should be
the project built against. This information is already defined in the
maven-android-plugin configuration, by the android.sdk.platform
property. This commit adds the ability to m2e-android to read this
property and generate the project.properties file automatically at
project import.