Use JDK8 in Travis
It has improved memory-management for example the whole PermGen space
has been replaced with the new MetaSpace. We thought maybe some OOM
causes our test to fail in Travis, so that's why i made this change,
but it did not help. This is still interesting, because it is a good
thing to run the build and the test on a newer VM.

#274