Update maven plugin versions
So they are working well even with JDK8. Precisely, the old version of
the findbugs plugin has issues with JDK8. Also added maven version
requirenment to the POM, so we can properly display newer versions with
versions:display-plugin-updates. Moreover, the build will fail if
someone tries to use a lower maven version.