Added support for Andmore.
Modified toolkit adaptor factory to create and use project adaptors for
ADT or Andmore depending on which Eclipse plugin is installed as part of
this change:

https://github.com/rgladwell/m2e-android/issues/294

Project adaptor creation logic was also spread out over several classes
so this change consolidated these branches into one path so that we can
ensure we always create the right toolkit adaptor.

Because the Andmore p2 repo points to a site updated nightly we have
marked Andmore builds as unstable. This is because each time a new
build occurs for Andmore the old update site and its timestamped
dependencies are deleted which breaks m2e-android builds until the
target dependencies can be manually updated.

Maybe there is a way to specify dependencies in Eclipse target files
without having to specify the exact version.