Add sourcepath provider for JUnit launch configs
The sources for non-runtime dependencies are already attached
automatically, but still not available in JUnit test debugging. This is
because the source lookup does not contain the non-runtime dependency
entries, only the run-time ones. This commit fixes this issue by adding
a new source path provider which decorates the standard m2e one, and
adds non-runtime dependencies too, to the JUnit source lookup path.

https://github.com/rgladwell/m2e-android/issues/188