Workaround Tycho issue with optional deps.
This commit resolves two issues:

Builds were failing due to issues with Tycho resolving optional OSGi
dependencies (required to support both Andmore and ADT in the same code
base).

A workaround as outlined here:

https://bugs.eclipse.org/bugs/show_bug.cgi?id=351842#c3

...resolved the issue. For some reason (possibly caching of build
targets in the local Maven repo?) this was missed during the original
commit.

As part of this fix, a new compilation target was also created so that
the Maven build has access to both Andmore and ADT APIs now required for
compilation.