Integrate with JRebel
JRebel is a tool which enables advanced hot code replacement in almost
every situations (when the traditional replacement technic fails),
thus removes the need of redeploying the plugin when debugging. This
commit adds the necessary VM argument to the run configurations, and
adds the rebel.xml file to the git ignore list. It does not effect
developers who do not use JRebel.