Fix Android container not exported in m2e 1.6
m2e 1.6 runs less configurator calls, which is a good thing, but has a
bad side effect for m2e-android: our configurator will be called
earlier, and the Android libraries container will not be found in the
classpath, because it is only added after m2e-android ran. We have to
add a classpath listener, this way we can be sure to be notified of the
addition of the container, and mark it as not exported.

https://github.com/rgladwell/m2e-android/issues/360
https://github.com/rgladwell/m2e-android/issues/353