Run after Java configurator in m2e 1.6
m2e adds a new 'runsAfter' directive, which let's project configurators
to define calling order. We use this directive to be sure to be called
after the Java configurator. However, due to a bug [1], this is not
working currently in m2e 1.6.0, so we are updating the targets to use
m2e 1.6.1 which has a fix for this.

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=471038

https://github.com/rgladwell/m2e-android/issues/359
https://github.com/rgladwell/m2e-android/issues/353